+++ CHANGELOG.md
@@ -1,3 +1,99 @@
+## [21.0.0](https://github.com/ndabAP/vue-command/compare/v20.0.0...v21.0.0) (2020-08-14)
+
+
+### ⚠ BREAKING CHANGES
+
+* fix autocompletion example
+
+### Documentation
+
+* fix autocompletion example ([fb20155](https://github.com/ndabAP/vue-command/commit/fb2015539ccb78e5301be218954253794c24f186))
+
+## [20.0.0](https://github.com/ndabAP/vue-command/compare/v19.0.0...v20.0.0) (2020-08-14)
+
+
+### ⚠ BREAKING CHANGES
+
+* remove unnecessary built-in property
+
+### Features
+
+* remove unnecessary built-in property ([294c5f7](https://github.com/ndabAP/vue-command/commit/294c5f757930876f1e6bb9a0f4b318302ea8fa46))
+
+## [19.0.0](https://github.com/ndabAP/vue-command/compare/v18.0.0...v19.0.0) (2020-08-14)
+
+
+### ⚠ BREAKING CHANGES
+
+* introduce built-in property
+* use built-in programs
+
+### Bug Fixes
+
+* introduce built-in property ([fd2d7ad](https://github.com/ndabAP/vue-command/commit/fd2d7ad9b08b6ead41c7542cd68ea4cdea07a9cf))
+* use built-in programs ([e9b8fc7](https://github.com/ndabAP/vue-command/commit/e9b8fc78e372ea2549bdd29e4512096e1f1f57d5))
+
+## [18.0.0](https://github.com/ndabAP/vue-command/compare/v17.0.0...v18.0.0) (2020-08-13)
+
+
+### ⚠ BREAKING CHANGES
+
+* add missing contributor hint
+
+### Documentation
+
+* add missing contributor hint ([a4c5c43](https://github.com/ndabAP/vue-command/commit/a4c5c43ee25245a64fcd2c4ba0c226613492b32a))
+
+## [17.0.0](https://github.com/ndabAP/vue-command/compare/v16.2.0...v17.0.0) (2020-08-11)
+
+
+### ⚠ BREAKING CHANGES
+
+* fix built-in commands
+
+### Documentation
+
+* fix built-in commands ([b87a4aa](https://github.com/ndabAP/vue-command/commit/b87a4aa88a8ad85cf6b8e71f2575ff3f7a2e5c74))
+
+## [16.2.0](https://github.com/ndabAP/vue-command/compare/v16.1.0...v16.2.0) (2020-08-11)
+
+
+### Features
+
+* hide title ([369638e](https://github.com/ndabAP/vue-command/commit/369638e24ae74d5da2823f861ccfee4cb336bdf6))
+
+## [16.1.0](https://github.com/ndabAP/vue-command/compare/v16.0.0...v16.1.0) (2020-08-11)
+
+
+### Features
+
+* make commands optional ([ac9e2b5](https://github.com/ndabAP/vue-command/commit/ac9e2b54b9a540b1919520461c3bdec7bd2171d2))
+
+## [16.0.0](https://github.com/ndabAP/vue-command/compare/v15.0.2...v16.0.0) (2020-08-11)
+
+
+### ⚠ BREAKING CHANGES
+
+* use not commands for built-in
+
+### Features
+
+* use not commands for built-in ([0371b81](https://github.com/ndabAP/vue-command/commit/0371b81a36d8c8beddceffdd55226372dc4b81ab))
+
+### [15.0.2](https://github.com/ndabAP/vue-command/compare/v15.0.1...v15.0.2) (2020-08-04)
+
+
+### Bug Fixes
+
+* allow to mutate stdin if not in progress ([f44ca73](https://github.com/ndabAP/vue-command/commit/f44ca73fe7d9c53eef173d66361a0bb519ff72bb))
+
+### [15.0.1](https://github.com/ndabAP/vue-command/compare/v15.0.0...v15.0.1) (2020-07-25)
+
+
+### Bug Fixes
+
+* accommodate for more when parsing ([20f9a8f](https://github.com/ndabAP/vue-command/commit/20f9a8fe7a36f8037edaf3a388354594c2d64e3a))
+
 ## [15.0.0](https://github.com/ndabAP/vue-command/compare/v14.0.0...v15.0.0) (2020-07-13)